+++ CHANGELOG.md
@@ -57,6 +57,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   name, where hash is base16 seahash value, and name is taken from the manifest.
   If it's not avaialble, then the previous naming scheme is used. Old file names
   will automatically be renamed to the new format if possible.
+- Launcher will load old game packages, but show a message about game package
+  being outdated.
 
 ### Removed